Refactor dao layer

+++ backend/src/main/java/com/codecool/quizzzz/controller/QuizController.java
@@ -26,29 +26,29 @@ public class QuizController {
   }
 
   @GetMapping("/{quizId}")
-  ResponseEntity<QuizDTO> getQuizById(@PathVariable int quizId) {
+  ResponseEntity<QuizDTO> getQuizById(@PathVariable Long quizId) {
     return ResponseEntity.ok().body(quizService.getById(quizId));
   }
 
   @PostMapping("/create")
-  ResponseEntity<Integer> createQuiz(@RequestBody NewQuizDTO newQuizDTO) {
-    int id = quizService.create(newQuizDTO);
+  ResponseEntity<Long> createQuiz(@RequestBody NewQuizDTO newQuizDTO) {
+    Long id = quizService.create(newQuizDTO);
     return ResponseEntity.created(URI.create(String.format("/quiz/%d", id))).body(id);
   }
 
   @PostMapping("/createempty")
-  ResponseEntity<Integer> createQuiz() {
-    int id = quizService.create();
+  ResponseEntity<Long> createQuiz() {
+    Long id = quizService.create();
     return ResponseEntity.created(URI.create(String.format("/quiz/%d", id))).body(id);
   }
 
   @DeleteMapping("/{quizId}")
-  ResponseEntity<Integer> deleteQuiz(@PathVariable int quizId) {
+  ResponseEntity<Long> deleteQuiz(@PathVariable Long quizId) {
     return ResponseEntity.ok().body(quizService.deleteById(quizId));
   }
 
   @PatchMapping("/{quizId}")
-  ResponseEntity<Integer> renameQuiz(@PathVariable int quizId, @RequestBody NewQuizDTO newQuizDTO) {
+  ResponseEntity<Long> renameQuiz(@PathVariable Long quizId, @RequestBody NewQuizDTO newQuizDTO) {
     return ResponseEntity.ok().body(quizService.rename(newQuizDTO, quizId));
   }
 }
